# swagger/portal-swagger.yaml
swagger: '2.0'
info:
  version: '0.9'
  title: Portal
  description: >-
    The API of the Portal (the UI). All of these API calls may return different results
    depending on the current session; you will need to login to get access to restricted
    resources, e.g. to APIs or Content. Most of these end points require an
    `accept: application/json` header to return machine readable results, otherwise you
    will be served the corresponding HTML page. The data you get back from these end points
    corresponds 1:1 with the data visible on the pages; in most cases it's just that the
    rendering has been left out.
...
paths:
...
  /admin/approvals:
    get:
      description: Retrieve a list of pending subscription approvals (admins only)
      tags:
        - admin
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed (only Admin)
  /admin/approvals/approve:
    post:
      description: Approve of a specific subscription (admin only)
      tags:
        - admin
      parameters:
        - name: body
          in: body
          description: Subscription to approve of
          schema:
            type: object
            properties:
              app:
                type: string
                description: Application ID to approve subscription for
              api:
                type: string
                description: API ID to approve subscription for
      responses:
        200:
          description: Successfully approved subscription
        403:
          description: Not allowed.
  /admin/approvals/decline:
    post:
      description: Decline a specific subscription (admin only)
      tags:
        - admin
      parameters:
        - name: body
          in: body
          description: Subscription to decline
          schema:
            type: object
            properties:
              app:
                type: string
                description: Application ID to approve subscription for
              api:
                type: string
                description: API ID to approve subscription for
      responses:
        204:
          description: Successfully declined subscription
        403:
          description: Not allowed
  /admin/users:
    get:
      description: Retrieve a list of all users (short info) (admin only)
      tags:
        - admin
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed (only Admin)
  /admin/applications:
    get:
      description: Retrive a list of all applications (admin only)
      tags:
        - admin
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed (only Admin)
  /admin/listeners:
    get:
      description: Retrieve a list of all webhook listeners (admin only)
      tags:
        - admin
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed (only Admin)
  /admin/listeners/{listenerId}:
    get:
      description: Retrieve information on a specific listener (admin only)
      tags:
        - admin
      parameters:
        - name: listenerId
          in: path
          type: string
          required: true
          description: The ID of the listener to retrieve information on
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed (only Admin)
  /admin/verifications:
    get:
      description: Retrieve information on pending verifications (email, password) (admin only)
      tags:
        - admin
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed (only Admin)
...